Add files via upload
Added CPU level 5.
As far as I can tell, it should never be able to lose on level 5. I didn't want to include it originally (what's the point of playing the game if no one can win), but finally added for completeness sake. If you want a winnable game, just select a lower CPU level.